ekf logs script: move GPS and EKF comparison script to seperate file
JIRA: PP-189
